+++ src/apis/private/index.ts
@@ -12,5 +12,8 @@ export { getChannelData } from "./channelData";
 // Video Data Scraping
 export { getVideoData } from "./videoData";
 
+// Featured Livestreams
+export { getFeaturedLivestreams } from "./featuredLivestreams";
+
 // Private Moderation Methods
 export { deleteMessage, setSlowMode } from "./moderation";